% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoReMiTraExplorer.R
\name{DoReMiTra_explorer}
\alias{DoReMiTra_explorer}
\title{Visualize radiation transcriptomic datasets in the form of SE objects from the DoReMiTra collection}
\usage{
DoReMiTra_explorer(se)
}
\arguments{
\item{se}{SummarizedExperiment object from the DoReMiTra collection}
}
\description{
Visualize radiation transcriptomic datasets in the form of SE objects from the DoReMiTra collection
}
\examples{
#se <- DoReMiTra::get_DoReMiTra_data("SE_Salah_2025_ExVivo", gene_symbol = TRUE)
#DoReMiTra_explorer(se)
}
